app/pyproject.toml
[project]
name = "fast_api_telegram"
version = "0.1.0"
description = ""
authors = [{ name = "peyman", email = "peyman0_1@yahoo.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-telegram-bot[job-queue,socks] (>=21.10,<22.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
